# Replace strings.Trim with strings.TrimFunc in ParseRDF